% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifVeg.R
\name{classifVeg}
\alias{classifVeg}
\title{Classify pixels into structural types}
\usage{
classifVeg(
  r_pca,
  k,
  method = c("kmeans", "gmm"),
  n_pca = 3,
  nstart = 25,
  spatial_weight = 0,
  coords = NULL
)
}
\arguments{
\item{r_pca}{PCA scores of pixels in structural space, as returned by
\code{PCALandscape()$r_pca}}

\item{k}{integer, number of groups (clusters)}

\item{method}{either "kmeans" or "gmm" (Gaussian Mixture Model)}

\item{n_pca}{number of PCA axes used in analysis}

\item{nstart}{integer, number of random sets for K-means (\code{method} = "kmeans"). Ignored if \code{method} = "gmm"}

\item{spatial_weight}{weight applied to spatial coordinates to encourage
contiguous patches. 0 ignores coordinates, >0 increases spatial grouping. Kmeans only}

\item{coords}{matrix or dataframe of X, Y coordinates corresponding to the rows of r_pca.
Required if spat_weight > 0.}
}
\value{
vector containing cluster assignments for each pixel
}
\description{
Classify pixels into structural types
}
