% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractPlotMetrics.R
\name{extractPlotMetrics}
\alias{extractPlotMetrics}
\title{Extract values from raster across polygons}
\usage{
extractPlotMetrics(r, p, fun = mean, ...)
}
\arguments{
\item{r}{SpatRaster object}

\item{p}{either an sf or \code{SpatVector} object containing plot polygons or
points, or a dataframe with two columns \code{x} and \code{y} describing coordinates in the same
coordinate system as \code{r}}

\item{fun}{function to summarize the extracted data by polygon geometry, see
\code{terra::extract()}}

\item{...}{additional parameters passed to \code{terra::extract()}}
}
\value{
dataframe containing extracted metrics for each polygon
}
\description{
Extract values from raster across polygons
}
