% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/footprint.R
\name{footprintMetrics}
\alias{footprintMetrics}
\title{Calculate mean structural metrics within every possible plot footprint}
\usage{
footprintMetrics(r, p_new_dim)
}
\arguments{
\item{r}{\code{SpatRaster} of structural metrics or PCA scores}

\item{p_new_dim}{dimensions of plots in the same coordinate system as \code{r}.
Either a single value for square plots, or a vector of two values for
rectangular plots.}
}
\value{
matrix with one row per complete footprint and one column per
layer in \code{r}
}
\description{
Calculate mean structural metrics within every possible plot footprint
}
\details{
Plots are compared to the landscape at the scale of a plot, rather than
the scale of a pixel. The landscape is represented by the mean values
within every possible plot footprint (a moving window), which can then
be compared to plots using e.g. \code{pcaDist()}. Footprints containing any
missing values are excluded.
}
