% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCALandscape.R
\name{PCALandscape}
\alias{PCALandscape}
\title{Run a PCA on landscape structural metrics, and optionally place plots in
that space}
\usage{
PCALandscape(r, p = NULL, ...)
}
\arguments{
\item{r}{SpatRaster with structural metrics}

\item{p}{optional dataframe containing structural metrics for each plot, e.g.
as returned by \code{extractPlotMetrics()}}

\item{...}{additional arguments passed to \code{prcomp()}}
}
\value{
list containing: \code{r_pca}: PCA object from pixel values, and
optionally \code{p_pca}: plot values in PCA space
}
\description{
Run a PCA on landscape structural metrics, and optionally place plots in
that space
}
