% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_stems}
\alias{sample_stems}
\title{Stem data from 400x400 m (16 ha) plot}
\format{
An object of class \code{data.frame} with 21247 rows and 6 columns.
}
\usage{
sample_stems
}
\description{
dataframe object of stem data from one plot.
}
\keyword{datasets}
