% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{san_lorenzo_plots}
\alias{san_lorenzo_plots}
\title{Plot locations in San Lorenzo}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 63 rows and 2 columns.
}
\usage{
san_lorenzo_plots
}
\description{
sf POLYGONS object of existing plots in San Lorenzo
}
\keyword{datasets}
