% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{san_lorenzo_rast}
\alias{san_lorenzo_rast}
\title{Forest structural metrics across San Lorenzo}
\format{
An object of class \code{PackedSpatRaster} of length 1.
}
\usage{
san_lorenzo_rast
}
\description{
terra raster object of forest structural metrics across San Lorenzo.
}
\details{
50x50 m spatial resolution
}
\keyword{datasets}
