% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{san_lorenzo_stems}
\alias{san_lorenzo_stems}
\title{Stem data from existing San Lorenzo plots}
\format{
An object of class \code{data.frame} with 996 rows and 5 columns.
}
\usage{
san_lorenzo_stems
}
\description{
dataframe object of stem data from existing San Lorenzo plots.
}
\keyword{datasets}
