% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siteOptiSample-package.R
\docType{package}
\name{siteOptiSample-package}
\alias{siteOptiSample}
\alias{siteOptiSample-package}
\title{siteOptiSample: Optimal Placement of Forest Plots in Structural Space}
\description{
Tools to assess how well existing forest plots represent the structural variation of a landscape, and to select locations for new plots that fill gaps in multidimensional structural space. Works with raster datasets (e.g. airborne LiDAR metrics) or tabular data, and provides several plot selection algorithms (meanmin, minimax, K-means, and Latin hypercube).
}
\author{
\strong{Maintainer}: John L. Godlee \email{johngodlee@gmail.com}

}
\keyword{internal}
