% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifRepres.R
\name{classifRepres}
\alias{classifRepres}
\title{Classify pixels by whether they are well-represented by existing (and
proposed) plots}
\usage{
classifRepres(r_pca, p, p_new = NULL, n_pca = 3, ci = 0.95)
}
\arguments{
\item{r_pca}{PCA scores of pixels in structural space, as returned by
\code{PCALandscape()$r_pca}}

\item{p}{PCA scores of candidate plots in structural space, as returned
by \code{PCALandscape()$p_pca}}

\item{p_new}{optional PCA scores of candidate plots in structural space, as
returned by \code{PCALandscape()$p_pca}}

\item{n_pca}{number of PCA axes used in analysis}

\item{ci}{quantile threshold for distances, e.g. 0.95 = 95th percentile of
distances among plots}
}
\description{
Classify pixels by whether they are well-represented by existing (and
proposed) plots
}
