% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanminSelect.R
\name{meanminSelect}
\alias{meanminSelect}
\title{Iteratively add candidate plots using the meanmin algorithm}
\usage{
meanminSelect(r_pca, p_pca, old_ind, new_ind, n_plots, p_new_dim, het_q)
}
\arguments{
\item{r_pca}{\code{SpatRaster} of PCA scores or raw structural metrics.}

\item{p_pca}{optional, PCA values or raw structural metrics of existing
plots in structural space}

\item{old_ind}{optional, pixel IDs in \code{r_pca} specifying existing plot locations.}

\item{new_ind}{optional, pixel IDs in \code{r_pca} specifying candidate plot locations.
If not supplied, the full set of valid pixels in \code{r_pca} is used.}

\item{n_plots}{maximum number of new plots to add.}

\item{p_new_dim}{dimensions of new plots in the same coordinate system as \code{r}.
A vector of two values. Should be perfectly divisible by the resolution
of \code{r_pca}.}

\item{het_q}{optional, numeric value between 0 and 1. Defines a threshold
to filter out internally heterogeneous candidate locations, preventing plots
being placed on sharp structural transitions. Represents maximum
quantile for total focal standard deviation of metrics within
candidate footprint (e.g., 0.8 excludes top 20\% most heterogeneous
areas).}
}
\value{
list of \code{sf} polygons for proposed new plots.
}
\description{
Iteratively add candidate plots using the meanmin algorithm
}
\details{
The mean-min algorithm aims to minimize the mean distance of all candidate
locations (pixels or rows) to their nearest plot in structural/PCA space.
It iteratively places new plots in locations that most reduce the average
location-to-nearest-plot distance. For each candidate, it computes how
adding a plot there would decrease the overall mean distance, choosing
the candidate that yields the largest decrease. As a result, selected plots
tend to concentrate in the most common/representative areas of structural space.
}
