% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanminSelect.R
\name{meanminSelect}
\alias{meanminSelect}
\title{Iteratively add candidate plots using the meanmin algorithm}
\usage{
meanminSelect(r_pca, p_pca, old_ind, new_ind, n_plots, p_new_dim)
}
\arguments{
\item{r_pca}{\code{SpatRaster}, \code{data.frame}, or \code{matrix} of PCA scores or raw
structural metrics. If spatial, values represent pixels across a landscape.
If non-spatial, each row represents a candidate plot}

\item{p_pca}{optional, PCA values or raw structural metrics of existing
plots in structural space}

\item{old_ind}{optional, pixel IDs (if spatial) or row indices (if non-spatial)
in \code{r_pca} specifying existing plot locations.}

\item{new_ind}{optiona, pixel IDs (if spatial) or row indices (if non-spatial)
in \code{r_pca} specifying candidate plot locations. If not supplied, the full
set of valid landscape pixels or dataframe rows in \code{r_pca} is used.}

\item{n_plots}{maximum number of new plots to add.}

\item{p_new_dim}{dimensions of new plots in the same coordinate system as \code{r}.
A vector of two values. Should be perfectly divisible by the resolution
of \code{r_pca}. \strong{Note:} Ignored if \code{r_pca} is a \code{data.frame} or \code{matrix}.}
}
\value{
If \code{r_pca} is a \code{SpatRaster}, returns a list of \code{sf} polygons for the
proposed new plots. If \code{r_pca} is a \code{data.frame} or \code{matrix}, returns a
numeric vector of row indices corresponding to the selected plots.
}
\description{
Iteratively add candidate plots using the meanmin algorithm
}
\details{
The mean-min algorithm aims to minimize the mean distance of all candidate
locations (pixels or rows) to their nearest plot in structural/PCA space.
It iteratively places new plots in locations that most reduce the average
location-to-nearest-plot distance. For each candidate, it computes how
adding a plot there would decrease the overall mean distance, choosing
the candidate that yields the largest decrease. As a result, selected plots
tend to concentrate in the most common/representative areas of structural space.
}
