% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanminSelect.R
\name{meanminSelect}
\alias{meanminSelect}
\title{Iteratively add candidate plots using the meanmin algorithm}
\usage{
meanminSelect(r_pca, p_pca, old_ind, new_ind, n_plots, p_new_dim)
}
\arguments{
\item{r_pca}{raster of PCA scores of pixels in structural space}

\item{p_pca}{PCA values of existing plots in structural space}

\item{old_ind}{optional, pixel IDs in \code{r_pca} specifying existing plot
locations.}

\item{new_ind}{optional, pixel IDs in \code{r_pca} specifying candidate plot
locations. If not supplied, the full set of landscape pixels in \code{r_pca}
is used.}

\item{n_plots}{maximum number of new plots to add}

\item{p_new_dim}{dimensions of new plots in the same coordinate system as
\code{r}. A vector of two values. Should be perfectly divisible by the
resolution of \code{r}.}
}
\value{
vector of pixel index values for proposed new plots
}
\description{
Iteratively add candidate plots using the meanmin algorithm
}
\details{
The meanmin algorithm aims to minimize the mean distance of all pixels to
their nearest plot in PCA space, iteratively placing new plots in pixels
that most reduce the average pixel-to-nearest-plot distance. For each
candidate pixel, compute how adding a plot to that pixel would decrease
the mean pixel-to-nearest-plot distance, then choose the pixel that
yields the largest decrease in mean distance. As a result, plots
concentrate in common areas structural space.
}
