% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimaxSelect.R
\name{minimaxSelect}
\alias{minimaxSelect}
\title{Iteratively add candidate plots using the minimax algorithm}
\usage{
minimaxSelect(r_pca, p_pca, old_ind, new_ind, n_plots, p_new_dim, het_q)
}
\arguments{
\item{r_pca}{\code{SpatRaster} of PCA scores or raw structural metrics.}

\item{p_pca}{optional, PCA values or raw structural metrics of existing
plots in structural space}

\item{old_ind}{optional, pixel IDs in \code{r_pca} specifying existing plot locations.}

\item{new_ind}{optional, pixel IDs in \code{r_pca} specifying candidate plot locations.
If not supplied, the full set of valid pixels in \code{r_pca} is used.}

\item{n_plots}{maximum number of new plots to add.}

\item{p_new_dim}{dimensions of new plots in the same coordinate system as \code{r}.
A vector of two values. Should be perfectly divisible by the resolution
of \code{r_pca}.}

\item{het_q}{optional, numeric value between 0 and 1. Defines a threshold
to filter out internally heterogeneous candidate locations, preventing plots
being placed on sharp structural transitions. Represents maximum
quantile for total focal standard deviation of metrics within
candidate footprint (e.g., 0.8 excludes top 20\% most heterogeneous
areas).}
}
\value{
list of \code{sf} polygons for proposed new plots.
}
\description{
Iteratively add candidate plots using the minimax algorithm
}
\details{
The minimax algorithm aims to minimise the maximum distance between
proposed plots and existing plots, iteratively placing new plots in
pixels with structural attributes most dissimilar to existing plots. For
each candidate pixel, it computes the distance to the nearest existing plot,
then chooses the pixel with the highest distance value. As a result, plots
occupy structural extremes.
}
