% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimaxSelect.R
\name{minimaxSelect}
\alias{minimaxSelect}
\title{Iteratively add candidate plots using the minimax algorithm}
\usage{
minimaxSelect(r_pca, p_pca, old_ind, new_ind, n_plots, p_new_dim)
}
\arguments{
\item{r_pca}{raster of PCA scores of pixels in structural space}

\item{p_pca}{PCA values of existing plots in structural space}

\item{old_ind}{optional, pixel IDs in \code{r_pca} specifying existing plot
locations.}

\item{new_ind}{optional, pixel IDs in \code{r_pca} specifying candidate plot
locations. If not supplied, the full set of landscape pixels in \code{r_pca}
is used.}

\item{n_plots}{maximum number of new plots to add}

\item{p_new_dim}{dimensions of new plots in the same coordinate system as
\code{r}. A vector of two values. Should be perfectly divisible by the
resolution of \code{r}.}
}
\value{
vector of pixel index values for proposed new plots
}
\description{
Iteratively add candidate plots using the minimax algorithm
}
\details{
The minimax algorithm aims to minimise the maximum distance between
proposed plots and existing plots, iteratively placing new plots in
pixels with structural attributes most dissimilar to existing plots. For
each candidate pixel, compute the distance to the nearest existing plot,
then choose the pixel with the highest distance value. As a result, plots
occupy structural extremes.
}
