% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaDist.R
\name{pcaDist}
\alias{pcaDist}
\title{Get nearest neighbour distances between two PCAs}
\usage{
pcaDist(x, y, w = NULL, n_pca = 3, k = 1, method = "euclidean")
}
\arguments{
\item{x}{PCA scores}

\item{y}{PCA scores}

\item{w}{optional, numeric vector of weights (one per column). Larger values
increase a variable’s contribution; if NULL, all variables are equally
weighted.}

\item{n_pca}{number of PCA axes used in analysis}

\item{k}{number of nearest neighbours in y to return for each row in x}

\item{method}{distance method, either "euclidean" or "mahalanobis"}
}
\value{
if \code{k = 1} a vector of distances to the nearest neighbour in \code{y} for
each row in \code{x}. If \code{k = 2} a matrix with k columns with ordered nearest
neighbour distances
}
\description{
Get nearest neighbour distances between two PCAs
}
