% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSelect.R
\name{plotSelect}
\alias{plotSelect}
\title{Recommend locations for plots}
\usage{
plotSelect(
  r,
  p = NULL,
  n_plots,
  p_new_dim = NULL,
  r_mask = NULL,
  pca = TRUE,
  n_pca = NULL,
  coord = NULL,
  het_q = NULL,
  method = meanminSelect
)
}
\arguments{
\item{r}{\code{SpatRaster} or dataframe with structural metrics}

\item{p}{optional, either an \code{sf} object containing polygons or points of
existing plots, a dataframe with two columns \code{x} and \code{y} describing
coordinates in the same coordinate system as \code{r}, or if \code{r} is not
spatial a vector of row indices in \code{r} specifying locations with
existing plots}

\item{n_plots}{maximum number of new plots to add}

\item{p_new_dim}{optional, dimensions of new plots in the same coordinate
system as \code{r}. Either a single value for square plots, or a vector of two
values for rectangular plots. Should be perfectly divisible by the
resolution of \code{r}}

\item{r_mask}{optional, a raster which defines a mask of potential plot
locations, e.g. based on accessibility. Ignored if \code{r} is not spatial}

\item{pca}{logical, if TRUE (default) the variables in \code{r} are run through a
PCA to reduce issues of collinearity among variables. If TRUE, provide
\code{n_pca}}

\item{n_pca}{optional, the number of PCA axes used to calculate the distance
between pixels and plots. If not provided and \code{pca} is TRUE, all PCA axes
are used}

\item{coord}{optional, if \code{r} is a dataframe, a character vector containing
two column names in \code{r} specifying the X and Y coordinates of the grid
cell centres. Coordinates must form a regular grid}

\item{het_q}{optional, numeric value between 0 and 1. Defines a threshold
to filter out internally heterogeneous candidate locations, preventing plots
being placed on sharp structural transitions. Represents maximum
quantile for total focal standard deviation of metrics within
candidate footprint (e.g., 0.8 excludes top 20\% most heterogeneous
areas)}

\item{method}{a function to optimally place plots, e.g. \code{meanminSelect()} or
\code{minimaxSelect()}}
}
\value{
sf dataframe with polygons of proposed new plots, or if \code{r} is a dataframe, a vector of row indices in \code{r} corresponding to the selected rows
}
\description{
Recommend locations for plots
}
